Feat: 트랜잭션에서 외부 API 호출 분리
